requirements.txt
pandas
numpy
scikit-learn
joblib
streamlit
flask
google-cloud-bigquery
google-cloud-storage
google-cloud-aiplatform
apache-beam[gcp]
lightgbm
torch
